!BF (Renderer) (CE-17308) Added render flag to set stencil for exclusion of terrain layer blending on objects !F (Sandbox) Added entity flag checkbox to set ignore terrain layer blending !F (Cvar) Added render debug cvar to control terrain layer blending !R (Sandbox) (3DEngine) Changed renderFlag to uint64 as defined in the renderer
Copied from Perforce
 Change: 1830246

+++ Code/Sandbox/EditorQt/Objects/EntityObject.cpp
@@ -2224,7 +2224,7 @@ void CEntityObject::SetSelected(bool bSelect)
 			IRenderNode* pRenderNode = m_pEntity->GetRenderNode();
 			if (pRenderNode)
 			{
-				int flags = pRenderNode->GetRndFlags();
+				uint64 flags = pRenderNode->GetRndFlags();
 				if (bSelect)
 				{
 					flags |= ERF_SELECTED;
